requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
pipenv==2021.5.29
platformdirs==2.3.0
psycopg2==2.7.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
virtualenv==20.7.2
virtualenv-clone==0.5.6
whitenoise==5.3.0
zipp==3.5.0
